separate getting events and generating summary
