+++ busy_beaver/apps/github_integration/summary/blocks.py
@@ -1,42 +1,24 @@
-from datetime import datetime, timedelta
+from datetime import datetime
 import logging
-from typing import List, NamedTuple
 
-from .summary import GitHubUserEvents
-from busy_beaver.models import GitHubSummaryUser
 from busy_beaver.toolbox.slack_block_kit import Divider, Section
 from busy_beaver.toolbox.slack_block_kit.elements import Image
 
 logger = logging.getLogger(__name__)
 
 
-class UserEvents(NamedTuple):
-    user: GitHubSummaryUser
-    events: GitHubUserEvents
-
-
 class GitHubSummaryPost:
-    def __init__(self, users: List[GitHubSummaryUser], boundary_dt: datetime):
-        self.users = users
-        self.boundary_dt = boundary_dt
-        self.now = boundary_dt + timedelta(days=1)
+    def __init__(self, all_user_events):
+        self.all_user_events = all_user_events
 
     def __repr__(self):  # pragma: no cover
         return "<GitHubSummaryPost>"
 
-    def create(self):
-        all_user_events = []
-        for idx, user in enumerate(self.users):
-            logger.info("Compiling stats for {0}".format(user))
-            user_events = GitHubUserEvents(user, self.boundary_dt)
-
-            if len(user_events) > 0:
-                all_user_events.append(UserEvents(user, user_events))
-
-        self.all_user_events = all_user_events
-
     def as_blocks(self):
-        output = [Section(f"*Daily GitHub Summary -- {self.now:%B %d, %Y}*"), Divider()]
+        output = [
+            Section(f"*Daily GitHub Summary -- {datetime.now():%B %d, %Y}*"),
+            Divider(),
+        ]
 
         if not self.all_user_events:
             output.append(Section(text="No activity to report."))
