+++ tests/apps/github_integration/summary/test_summary.py
@@ -2,7 +2,7 @@ from datetime import timedelta
 
 import pytest
 
-from busy_beaver.apps.github_integration.summary.summary import GitHubUserEvents
+from busy_beaver.apps.github_integration.summary.workflow import GitHubUserEvents
 from busy_beaver.toolbox import utc_now_minus
